refactor(EntityModel): move configuration extension to separate file
